+++ restvolley/src/main/java/com/hujiang/restvolley/download/RestVolleyDownload.java
@@ -67,6 +67,7 @@ public class RestVolleyDownload {
         mRequestBuilder = new Request.Builder();
         mRequestBuilder.get();
         mOkHttpClient = requestEngine.okHttpClient;
+        mOkHttpClient.setFollowRedirects(true);
     }
 
     /**
@@ -169,7 +170,6 @@ public class RestVolleyDownload {
      */
     public RestVolleyDownload setWriteTimeout(long timeout) {
         mOkHttpClient.setWriteTimeout(timeout, TimeUnit.MILLISECONDS);
-
         return this;
     }
 
@@ -255,14 +255,18 @@ public class RestVolleyDownload {
 
             @Override
             public void onResponse(Response response) throws IOException {
-                //new download file
-                final File localFile = newFile(localPath + SUFIX_TMP);
-                //write stream
-                if (response.isSuccessful() && writeStream2File(response, localFile, mIsAppend, listener)) {
-                    notifyDownloadSuccess(request.urlString(), new File(localFile.getAbsolutePath().substring(0, localFile.getAbsolutePath().indexOf(SUFIX_TMP)))
-                            , response.code(), response.headers(), listener);
-                } else {
-                    notifyDownloadError(request.urlString(), new Exception(""), response.code(), response.headers(), listener);
+                try {
+                    //new download file
+                    final File localFile = newFile(localPath + SUFIX_TMP);
+                    //write stream
+                    if (response.isSuccessful() && writeStream2File(response, localFile, mIsAppend, listener)) {
+                        notifyDownloadSuccess(request.urlString(), new File(localFile.getAbsolutePath().substring(0, localFile.getAbsolutePath().indexOf(SUFIX_TMP)))
+                                , response.code(), response.headers(), listener);
+                    } else {
+                        notifyDownloadError(request.urlString(), new Exception(""), response.code(), response.headers(), listener);
+                    }
+                } catch (Exception e) {
+                   notifyDownloadError(request.urlString(), e, response.code(), response.headers(), listener);
                 }
             }
         });