+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.0.9(2016-11-28)
+* fix: no onDownloadFailure callback when SDCard is full
+
 ### v1.0.8(2016-11-07)
 * add method RestVolleyImageLoader#getDiskCacheDir(), RestVolleyImageLoader#getDiskCachePath()
 * add method RestVolleyImageLoader#displayImage(final String uri, final ImageView imageView, final ImageDisplayer imageDisplayer)